Extract core logics to the `fastmsa`  module
