--- tests/app/apps/fastapi.py
@@ -1,4 +1,0 @@
-"""`FastAPI <https://fastapi.tiangolo.com/>`_ 로 구현된 REST API 앱."""
-from fastapi import FastAPI
-
-app = FastAPI()
