+++ tests/app/services/__init__.py
@@ -1,5 +1,1 @@
-"""서비스 모듈.
-
-`from . import services` 로 하위 서비스들을 한 번에 로드될 수 있도록 합니다.
-"""
-from tests.app.services import batch
+from . import batch
