+++ tests/e2e/test_api.py
@@ -5,25 +5,27 @@ from typing import Callable, Optional
 import pytest
 import requests
 
-from tests.app import config
+from fastmsa import FastMsa
 
 from tests import random_batchref, random_orderid, random_sku
 
 AddStockFunc = Callable[[list[tuple[str, str, int, Optional[str]]]], None]
 
 
-def post_to_add_batch(ref: str, sku: str, qty: int, eta: Optional[str]) -> None:
+def post_to_add_batch(
+    msa: FastMsa, ref: str, sku: str, qty: int, eta: Optional[str]
+) -> None:
     """서비스 엔드포인트 `POST /batches` 를 통해 배치를 추가합니다."""
-    url = config.get_api_url()
+    url = msa.config.get_api_url()
     r = requests.post(
         f"{url}/batches", json={"ref": ref, "sku": sku, "qty": qty, "eta": eta}
     )
     assert r.status_code == 201
 
 
-def delete_batches(sku: str, refs: list[str]) -> None:
+def delete_batches(msa: FastMsa, sku: str, refs: list[str]) -> None:
     """서비스 엔드포인트 `DELETE /batches` 를 통해 배치를 삭제합니다."""
-    url = config.get_api_url()
+    url = msa.config.get_api_url()
     res = requests.delete(
         f"{url}/batches",
         json={
@@ -35,7 +37,7 @@ def delete_batches(sku: str, refs: list[str]) -> None:
 
 
 @pytest.fixture
-def setup_batches():
+def setup_batches(msa: FastMsa):
     _batch_refs = []
     _sku = ""
 
@@ -48,22 +50,23 @@ def setup_batches():
 
     yield wrapper
 
-    delete_batches(_sku, _batch_refs)
+    delete_batches(msa, _sku, _batch_refs)
 
 
 @pytest.mark.usefixtures("server")
 def test_happy_path_returns_201_and_allocated_batch(
     # pylint: disable=redefined-outer-name
-    setup_batches: Callable[[str, int], list[str]]
+    msa: FastMsa,
+    setup_batches: Callable[[str, int], list[str]],
 ) -> None:
     sku, othersku = random_sku(), random_sku("other")
     earlybatch, laterbatch = setup_batches(sku, 2)
     [otherbatch] = setup_batches(othersku, 1)
-    post_to_add_batch(laterbatch, sku, 100, "2021-01-02")
-    post_to_add_batch(earlybatch, sku, 100, "2021-01-01")
-    post_to_add_batch(otherbatch, othersku, 100, None)
+    post_to_add_batch(msa, laterbatch, sku, 100, "2021-01-02")
+    post_to_add_batch(msa, earlybatch, sku, 100, "2021-01-01")
+    post_to_add_batch(msa, otherbatch, othersku, 100, None)
     data = {"orderid": random_orderid(), "sku": sku, "qty": 3}
-    url = config.get_api_url()
+    url = msa.config.get_api_url()
     res = requests.post(f"{url}/batches/allocate", json=data)
     assert res.status_code == 201
     assert res.json()["batchref"] == earlybatch
